Chunk out ignore-repos
